+++ Expanded Cheat Menu/Changelog.md
@@ -11,6 +11,7 @@
 ### Fixed:
 - Issue with the cheat menu in the main menu (menu items using UICity to display values).
 - Buildings>*-free Building issue (thanks OyeBhotka).
+- Rocket>Change Resupply Settings didn't work unless you used Load Game (thanks OyeBhotka).
 
 ## v10.2 (24 Oct 2018)
 ### Added: